pyproject.toml
[project]
name = "framereader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "bs4>=0.0.2",
    "cryptography>=45.0.4",
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "lap>=0.5.12",
    "matplotlib>=3.10.3",
    "opencv-python>=4.11.0.86",
    "pydantic-settings>=2.9.1",
    "pyjwt==2.8.0",
    "pymysql>=1.1.1",
    "pytest>=8.4.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
    "uvicorn>=0.34.3",
    "websockets>=15.0.1",
]
